Use styles from new macaw, extract utils function
